added sweetAlert

+++ src/app/components/book/book.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, TemplateRef } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
+import  Swal  from 'sweetalert2';
 
 import { Book } from '../../models/book.model';
 import { BookService } from '../../service/book.service';
@@ -21,18 +22,24 @@ export class BookComponent implements OnInit {
   idAtual: string | undefined;
 
   _bookFilter!: string;
+
+  booksFiltrados: any = [];
+
   get bookFilter(): string {
     return this._bookFilter;
   }
+
   set bookFilter(value: string) {
     this._bookFilter = value;
     this.booksFiltrados = this.bookFilter ? this.filtrarBooks(this.bookFilter) : this.books;
   }
 
-  booksFiltrados: any = [];
-
   constructor(private bookService: BookService, private modalService: BsModalService) {
     this.books = [];
+  }
+
+  ngOnInit(): void {
+    this.getBooks();
   }
 
   openModalForm(form: TemplateRef<any>, id: string | undefined) {
@@ -56,10 +63,6 @@ export class BookComponent implements OnInit {
       this.bookInfo = undefined
 
     console.log(id, this.bookInfo);
-  }
-
-  ngOnInit(): void {
-    this.getBooks();
   }
 
   filtrarBooks(filtrarPor: string): any {
@@ -100,11 +103,21 @@ export class BookComponent implements OnInit {
     if (!this.idAtual) { // Adicionar
       this.bookService.addBook(myForm.value).subscribe(
         (book: Book) => {
+          Swal.fire({
+            title: 'Salvo!',
+            text: 'O livro foi adicionado com sucesso.',
+            icon: 'success',
+            showConfirmButton: false,
+            timer: 1500
+          })
           this.getBooks();
-          console.log('Adicionado com Sucesso!');
         },
         (erro: any) => {
-          alert("Erro ao salvar os dados!");
+          Swal.fire({
+            title: 'Erro!',
+            text: 'Erro ao adicionar livro.',
+            icon: 'error'
+          })
           console.log(erro);
         }
       );
@@ -112,11 +125,21 @@ export class BookComponent implements OnInit {
       myForm.value.id = this.idAtual;
       this.bookService.updateBook(this.idAtual, myForm.value).subscribe(
         (book: Book) => {
+          Swal.fire({
+            title: 'Salvo!',
+            text: 'O livro foi editado com sucesso.',
+            icon: 'success',
+            showConfirmButton: false,
+            timer: 1500
+          })
           this.getBooks();
-          console.log('Editado com Sucesso!');
         },
         (erro: any) => {
-          alert("Erro ao salvar os dados!");
+          Swal.fire({
+            title: 'Erro!',
+            text: 'Erro ao editar livro.',
+            icon: 'error'
+          })
           console.log(erro);
         }
       );
@@ -125,14 +148,38 @@ export class BookComponent implements OnInit {
   }
 
   deleteBook(id: string) {
-    this.bookService.deleteBook(id).subscribe(
-      (message: string) => {
-        this.getBooks();
-        console.log('Removido com Sucesso!');
-      },
-      (erro: any) => {
-        console.log(erro);
+    Swal.fire({
+      title: 'Você tem certeza?',
+      text: "O livro será excluido!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      confirmButtonText: 'Excluir',
+      cancelButtonText: 'Cancelar',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.bookService.deleteBook(id).subscribe(
+          (message: string) => {
+            Swal.fire({
+              title: 'Excluido!',
+              text: 'O livro foi excluido com sucesso.',
+              icon: 'success',
+              showConfirmButton: false,
+              timer: 1500
+            })
+            this.getBooks();
+          },
+          (erro: any) => {
+            Swal.fire({
+              title: 'Erro!',
+              text: 'Erro ao excluir livro.',
+              icon: 'error'
+            })
+            console.log(erro);
+          }
+        );
       }
-    );
+    })
   }
 }
